requirements/common.txt
Babel==1.3
Flask-Babel==0.9
Flask-Bootstrap==3.0.3.1
Flask-HTTPAuth==2.2.0
Flask-Login==0.2.7
Flask-Mail==0.9.0
Flask-Migrate==1.1.0
Flask-Moment==0.2.1
Flask-PageDown==0.1.4
Flask-Redis==0.0.5
Flask-SQLAlchemy==1.0
Flask-Script==0.6.6
Flask-WTF==0.9.4
Flask==0.10.1
GitPython==0.1.7
Jinja2==2.7.1
Mako==0.9.1
Markdown==2.3.1
MarkupSafe==0.18
SQLAlchemy==0.8.4
WTForms==1.0.5
Werkzeug==0.9.4
alembic==0.6.2
async==0.6.1
bleach==1.4.0
blinker==1.3
gitdb==0.5.4
hiredis==0.1.4
html5lib==1.0b3
itsdangerous==0.23
pyapi-gitlab==6.2.3
redis==2.10.3
requests==2.1.0
sh==1.09
six==1.4.1
smmap==0.8.2
